feat(graph): graph view is now clickable and nodes can be dragged on their own
Added two interactions with the graph when touching a node.

If the press is short,
a certain function is triggered (e.g., display the profile of clicked user).

If the press is longer than a certain duration (set to 100ms),
then the node becomes draggable until you remove your finger from the screen.